Changed root namespace to com.genexus.gxserver.client
