+++ src/jaxws-bindings/CommonHandler.xml
@@ -10,8 +10,8 @@
         <jws:handler-chains>
             <jws:handler-chain>
                 <jws:handler>
-                    <jws:handler-name>genexus.server.clients.common.SOAPHeaderHandler</jws:handler-name>
-                    <jws:handler-class>com.genexus.server.clients.common.SOAPHeaderHandler</jws:handler-class>
+                    <jws:handler-name>genexus.gxserver.client.clients.common.SOAPHeaderHandler</jws:handler-name>
+                    <jws:handler-class>com.genexus.gxserver.client.clients.common.SOAPHeaderHandler</jws:handler-class>
                 </jws:handler>
             </jws:handler-chain>
         </jws:handler-chains>
@@ -26,12 +26,12 @@
     <!-- We should use these same bindings for all services so they share the generated packages -->
     <jaxb:bindings node="//xsd:schema[@targetNamespace='http://schemas.microsoft.com/2003/10/Serialization/']" schemaLocation="xs:http://schemas.microsoft.com/2003/10/Serialization/">
         <jaxb:schemaBindings>
-            <jaxb:package name="com.genexus.server.services.serialization"/>
+            <jaxb:package name="com.genexus.gxserver.client.services.serialization"/>
         </jaxb:schemaBindings>
     </jaxb:bindings>
     <jaxb:bindings node="//xsd:schema[@targetNamespace='http://schemas.datacontract.org/2004/07/GeneXus.Server.Contracts']" schemaLocation="xs:http://schemas.datacontract.org/2004/07/GeneXus.Server.Contracts">
         <jaxb:schemaBindings>
-            <jaxb:package name="com.genexus.server.services.contracts"/>
+            <jaxb:package name="com.genexus.gxserver.client.services.contracts"/>
         </jaxb:schemaBindings>
     </jaxb:bindings>
