+++ src/jaxws-bindings/HelperService.svc.singlewsdl_handler.xml
@@ -8,19 +8,19 @@
 
     <jaxb:bindings node="wsdl:definitions">
         <!-- Classes specific to this service (HelperService) use this package name -->
-        <jaxws:package name="com.genexus.server.services.helper"/>
+        <jaxws:package name="com.genexus.gxserver.client.services.helper"/>
     </jaxb:bindings>
 
     <!-- Set packange name binding for our specicic schema -->
     <jaxb:bindings node="//xsd:schema[@targetNamespace='http://tempuri.org/']" schemaLocation="xs:http://tempuri.org/">
         <jaxb:schemaBindings>
-            <jaxb:package name="com.genexus.server.services.helper"/>
+            <jaxb:package name="com.genexus.gxserver.client.services.helper"/>
         </jaxb:schemaBindings>
     </jaxb:bindings>
 
     <jaxb:bindings node="//xsd:schema[@targetNamespace='http://schemas.microsoft.com/2003/10/Serialization/Arrays']" schemaLocation="xs:http://schemas.microsoft.com/2003/10/Serialization/Arrays">
         <jaxb:schemaBindings>
-            <jaxb:package name="com.genexus.server.services.serialization.arrays"/>
+            <jaxb:package name="com.genexus.gxserver.client.services.serialization.arrays"/>
         </jaxb:schemaBindings>
     </jaxb:bindings>
