+++ src/main/java/com/genexus/gxserver/client/clients/TeamWorkService2Client.java
@@ -21,26 +21,26 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.genexus.server.clients;
+package com.genexus.gxserver.client.clients;
 
-import com.genexus.helpers.XmlHelper;
-import com.genexus.server.clients.common.ServiceData;
-import com.genexus.server.clients.common.ServiceInfo;
-import com.genexus.server.clients.common.TransferPropConstants;
-import com.genexus.server.clients.common.TransferPropHelper;
-import com.genexus.server.clients.common.WithLocalContextClassLoader;
-import com.genexus.server.info.KBList;
-import com.genexus.server.info.RevisionList;
-import com.genexus.server.info.VersionList;
-import com.genexus.server.services.contracts.ArrayOfServerMessage;
-import com.genexus.server.services.contracts.ArrayOfTransferProp;
-import com.genexus.server.services.teamwork.FileTransfer;
-import com.genexus.server.services.teamwork.ITeamWorkService2;
-import com.genexus.server.services.teamwork.ITeamWorkService2GetRevisionsGXServerExceptionFaultFaultMessage;
-import com.genexus.server.services.teamwork.ITeamWorkService2GetVersionsGXServerExceptionFaultFaultMessage;
-import com.genexus.server.services.teamwork.ITeamWorkService2HostedKBsGXServerExceptionFaultFaultMessage;
-import com.genexus.server.services.teamwork.SimpleTransfer;
-import com.genexus.server.services.teamwork.TeamWorkService2;
+import com.genexus.gxserver.client.helpers.XmlHelper;
+import com.genexus.gxserver.client.clients.common.ServiceData;
+import com.genexus.gxserver.client.clients.common.ServiceInfo;
+import com.genexus.gxserver.client.clients.common.TransferPropConstants;
+import com.genexus.gxserver.client.clients.common.TransferPropHelper;
+import com.genexus.gxserver.client.clients.common.WithLocalContextClassLoader;
+import com.genexus.gxserver.client.info.KBList;
+import com.genexus.gxserver.client.info.RevisionList;
+import com.genexus.gxserver.client.info.VersionList;
+import com.genexus.gxserver.client.services.contracts.ArrayOfServerMessage;
+import com.genexus.gxserver.client.services.contracts.ArrayOfTransferProp;
+import com.genexus.gxserver.client.services.teamwork.FileTransfer;
+import com.genexus.gxserver.client.services.teamwork.ITeamWorkService2;
+import com.genexus.gxserver.client.services.teamwork.ITeamWorkService2GetRevisionsGXServerExceptionFaultFaultMessage;
+import com.genexus.gxserver.client.services.teamwork.ITeamWorkService2GetVersionsGXServerExceptionFaultFaultMessage;
+import com.genexus.gxserver.client.services.teamwork.ITeamWorkService2HostedKBsGXServerExceptionFaultFaultMessage;
+import com.genexus.gxserver.client.services.teamwork.SimpleTransfer;
+import com.genexus.gxserver.client.services.teamwork.TeamWorkService2;
 import jakarta.xml.bind.JAXBException;
 import jakarta.xml.ws.BindingProvider;
 import jakarta.xml.ws.Holder;
