+++ src/main/java/com/genexus/gxserver/client/clients/common/ServiceData.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.genexus.server.clients.common;
+package com.genexus.gxserver.client.clients.common;
 
 import java.net.MalformedURLException;
 import java.net.URL;
