+++ src/main/java/com/genexus/gxserver/client/clients/common/WithLocalContextClassLoader.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2020 GeneXus S.A..
+ * Copyright 2021 GeneXus S.A..
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,29 +21,34 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.genexus.server.clients.common;
+package com.genexus.gxserver.client.clients.common;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import com.genexus.gxserver.client.helpers.ThrowingSupplier;
 
 /**
+ * Defines a critical section during which the context class loader is set to
+ * one obtained from the current class and thus being able to find classes
+ * contained in this package or those on which it depends.
+ *
+ * This fixes class loading errors when accessing web services using JAX-WS.
+ * More info about the problem at
+ * https://www.eclipse.org/forums/index.php/t/266362/
  *
  * @author jlr
  */
-public class ServiceInfo {
+public class WithLocalContextClassLoader {
 
-    public final String secureIdentifier;
-    public final String nonSecureIdentifier;
+    public static <T, E extends Throwable> T call(ThrowingSupplier<T, E> s) throws E {
+        T result = null;
 
-    @SuppressFBWarnings("URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD")
-    public final String secureBindingName;
+        ClassLoader tccl = Thread.currentThread().getContextClassLoader();
+        Thread.currentThread().setContextClassLoader(WithLocalContextClassLoader.class.getClassLoader());
+        try {
+            result = s.get();
+        } finally {
+            Thread.currentThread().setContextClassLoader(tccl);
+        }
 
-    @SuppressFBWarnings("URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD")
-    public final String nonSecureBindingName;
-
-    public ServiceInfo(String secureIdentifier, String nonSecureIdentifier, String secureBindingName, String nonSecureBindingName) {
-        this.secureIdentifier = secureIdentifier;
-        this.nonSecureIdentifier = nonSecureIdentifier;
-        this.secureBindingName = secureBindingName;
-        this.nonSecureBindingName = nonSecureBindingName;
+        return result;
     }
 }
