+++ src/main/java/com/genexus/gxserver/client/helpers/TrueOrNothingBooleanAdapter.java
@@ -21,31 +21,32 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.genexus.server.info;
+package com.genexus.gxserver.client.helpers;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import jakarta.xml.bind.annotation.adapters.XmlAdapter;
 
 /**
  *
  * @author jlr
+ *
+ * For TeamWorkService2.GetVersions() GXserver exports boolean elements only if
+ * they are true, and it uses "True" (title case) if that's the case.
+ *
+ * This adapter checks string values ignoring case (so as to recognize "True")
+ * and converts to null when the value is false (so as to avoid writing it).
+ *
+ * This makes sure the roundtrip (converting the received XML string to Document
+ * and back to string) returns the same result.
  */
-public class ServerInfo {
+public class TrueOrNothingBooleanAdapter extends XmlAdapter<String, Boolean> {
 
-    @SuppressFBWarnings( "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" )
-    public String serverVersion = "";
+    @Override
+    public Boolean unmarshal(String strValue) throws Exception {
+        return strValue.compareToIgnoreCase("true") == 0;
+    }
 
-    @SuppressFBWarnings( "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" )
-    public Boolean isAvailable = false;
-
-    @SuppressFBWarnings( "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" )
-    public Boolean isSecure = false;
-
-    @SuppressFBWarnings( "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" )
-    public Boolean supportsTokenAuthentication = false;
-
-    @SuppressFBWarnings( "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" )
-    public Boolean allowsGXtest = false;
-
-    @SuppressFBWarnings( "URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD" )
-    public Boolean allowsCustomBinding = false;
+    @Override
+    public String marshal(Boolean booleanValue) throws Exception {
+        return booleanValue ? "True" : null;
+    }
 }
