+++ src/spotbugs/excludesFilter.xml
@@ -6,21 +6,21 @@
 
     <!-- Code generated by jaxws-wsimport -->
     <Match>
-        <Class name="com.genexus.server.services.contracts.GXServerException" />
+        <Class name="com.genexus.gxserver.client.services.contracts.GXServerException" />
         <Bug pattern="NM_CLASS_NOT_EXCEPTION" />
     </Match>
     <Match>
-        <Class name="com.genexus.server.services.teamwork.FileTransfer" />
+        <Class name="com.genexus.gxserver.client.services.teamwork.FileTransfer" />
         <Method name="getFileByteStream" />
         <Bug pattern="EI_EXPOSE_REP" />
     </Match>
     <Match>
-        <Class name="com.genexus.server.services.teamwork.FileTransfer" />
+        <Class name="com.genexus.gxserver.client.services.teamwork.FileTransfer" />
         <Method name="setFileByteStream" />
         <Bug pattern="EI_EXPOSE_REP2" />
     </Match>
     <Match>
-        <Class name="~.*\.genexus\.server\.services.*" />
+        <Class name="~.*\.genexus\.gxserver\.client\.services.*" />
         <Field name="faultInfo" />
         <Bug pattern="SE_BAD_FIELD" />
     </Match>
